gemini/requirements.txt
google-genai>=0.2.0
Pillow>=10.0.0
